+++ src/main/java/org/wise/portal/presentation/web/controllers/admin/SiteController.java
@@ -43,7 +43,7 @@ import javax.servlet.http.HttpServletRequest;
 @RequestMapping("/site")
 public class SiteController {
 
-  @RequestMapping(value = {"", "/student", "/student/**", "/teacher", "/teacher/**", "/login", "/join", "/join/**", "/news", "/about", "/features", "/privacy"},
+  @RequestMapping(value = {"", "/student", "/student/**", "/teacher", "/teacher/**", "/login", "/login/**", "/join", "/join/**", "/news", "/about", "/features", "/privacy"},
       method = RequestMethod.GET)
   protected String showSite() {
     return "forward:/site/index.html";
