+++ src/main/webapp/site/src/app/login/login-google-user-not-found/login-google-user-not-found.component.html
@@ -1,0 +1,9 @@
+<mat-card class="standalone__content standalone__content--sm mat-elevation-z4">
+  <mat-card-content>
+    Your Google User ID was not found in WISE.
+  </mat-card-content>
+  <mat-divider></mat-divider>
+  <mat-card-actions>
+    <p>New to WISE? <a routerLink="/join">Join for free!</a></p>
+  </mat-card-actions>
+</mat-card>
