+++ src/main/webapp/site/src/app/login/login-google-user-not-found/login-google-user-not-found.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-login-google-user-not-found',
+  templateUrl: './login-google-user-not-found.component.html',
+  styleUrls: ['./login-google-user-not-found.component.scss']
+})
+export class LoginGoogleUserNotFoundComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
