+++ src/main/webapp/site/src/app/login/login-home/login-home.component.html
@@ -1,0 +1,59 @@
+<mat-card class="standalone__content standalone__content--sm mat-elevation-z4">
+  <mat-card-content>
+    <form role="form" (submit)="login()">
+      <h2 class="standalone__title accent">Sign in to WISE</h2>
+      <p class="warn center standalone__error inactive"
+         [ngClass]="{'active': error}">
+        Username and password not recognized. Please try again.
+      </p>
+      <p>
+        <mat-form-field appearance="fill" fxFlex>
+          <mat-label>Username</mat-label>
+          <input matInput
+                 id="username"
+                 name="username"
+                 [(ngModel)]="credentials.username"
+                 [disabled]="processing"
+                 autofocus />
+        </mat-form-field>
+      </p>
+      <p>
+        <mat-form-field appearance="fill" fxFlex>
+          <mat-label>Password</mat-label>
+          <input matInput
+                 id="password"
+                 name="password"
+                 type="password"
+                 [disabled]="processing"
+                 [(ngModel)]="credentials.password" />
+        </mat-form-field>
+      </p>
+      <p>
+        <button fxFlex
+                mat-raised-button
+                color="primary"
+                type="submit"
+                [disabled]="processing">
+          <span *ngIf="!processing">Sign In</span>
+          <span *ngIf="processing">Connecting...</span>
+        </button>
+      </p>
+    </form>
+    <span *ngIf="isGoogleAuthenticationEnabled">
+      <p class="center mat-subheading-1">- or -</p>
+      <p class="center">
+        <button class="button--social-login button--google"
+                color="accent"
+                mat-raised-button
+                (click)="socialSignIn('google')">
+          <img src="assets/img/icons/g-logo.png" alt="Google logo" />Sign in with Google
+        </button>
+      </p>
+    </span>
+  </mat-card-content>
+  <mat-divider></mat-divider>
+  <mat-card-actions>
+    <p><a routerLink="/lost">Forgot username or password?</a></p>
+    <p>New to WISE? <a routerLink="/join">Join for free!</a></p>
+  </mat-card-actions>
+</mat-card>
