+++ src/main/webapp/site/src/app/login/login-routing.module.ts
@@ -2,12 +2,17 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 import { LoginComponent } from "./login.component";
+import { LoginHomeComponent } from "./login-home/login-home.component";
+import { LoginGoogleUserNotFoundComponent } from "./login-google-user-not-found/login-google-user-not-found.component";
 
 const loginRoutes: Routes = [
   {
     path: '',
     component: LoginComponent,
-    children: []
+    children: [
+      { path: '', component: LoginHomeComponent },
+      { path: 'login/googleUserNotFound', component: LoginGoogleUserNotFoundComponent }
+    ]
   }
 ];
