+++ src/main/webapp/site/src/app/login/login.module.ts
@@ -1,10 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { LoginComponent } from "./login.component";
+import { LoginComponent } from './login.component';
+import { LoginGoogleUserNotFoundComponent } from './login-google-user-not-found/login-google-user-not-found.component';
+import { LoginHomeComponent } from "./login-home/login-home.component";
+import { LoginRoutingModule } from "./login-routing.module";
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
-import { LoginRoutingModule } from "./login-routing.module";
 
 import {
   MatButtonModule,
@@ -32,6 +34,9 @@ const materialModules = [
     materialModules
   ],
   declarations: [
+    LoginComponent,
+    LoginHomeComponent,
+    LoginGoogleUserNotFoundComponent,
     LoginComponent
   ],
   exports: [
